backend/requirements.txt
# PDF Processing
PyPDF2==3.0.1
PyMuPDF==1.26.3
unstructured[pdf]==0.18.3
pillow==11.3.0
pdf2image>=1.16.0
pdfplumber>=0.9.0
pytesseract>=0.3.10

# Vector Database & Embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Azure OpenAI / OpenAI
openai>=1.0.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.65.0
pandas>=2.0.0
numpy>=1.24.0
rich>=14.0.0

# FastAPI & Web Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0
pydantic>=2.0.0